requirements.txt
asttokens==2.0.5
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
DateTime==4.4
debugpy==1.6.0
decorator==5.1.1
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.31.2
google-auth==2.6.2
google-auth-oauthlib==0.5.1
gspread==5.2.0
idna==3.3
imbalanced-learn==0.9.0
imblearn==0.0
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Sastrawi==1.0.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
stack-data==0.2.0
statistics==1.0.3.5
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
wincertstore==0.2
wordcloud==1.8.1
zope.interface==5.4.0
